Make DefaultDistanceCalculationStrategyTest a roboelectric test #STAD-35
