+++ datacapturing/build.gradle
@@ -116,6 +116,5 @@ dependencies {
     testImplementation "org.mockito:mockito-core:$rootProject.ext.mockitoVersion"
     // Optional - For better debuggable asserts
     testImplementation "org.hamcrest:hamcrest-all:$rootProject.ext.hamcrestVersion"
-    testImplementation project(":testutils")
     testImplementation "org.robolectric:robolectric:$rootProject.ext.robolectricVersion"
 }
